feat: add route

+++ routes/api.php
@@ -20,30 +20,39 @@ use Illuminate\Support\Facades\Route;
 */
 
 Route::group(['middleware' => 'api', 'prefix' => 'auth'], function ($router) {
-    Route::post('login', [AuthController::class, 'login']);
-    Route::post('logout', [AuthController::class, 'logout']);
-    Route::get('checkUser', [AuthController::class, 'checkUser']);
+    Route::post('login', [AuthController::class, 'login']); // kelurahan, posyandu, puskesmas
+    Route::post('logout', [AuthController::class, 'logout']); // kelurahan, posyandu, puskesmas
+    Route::get('checkUser', [AuthController::class, 'checkUser']); // kelurahan, posyandu, puskesmas
 });
 
 Route::group(['middleware' => 'api', 'prefix' => 'posyandu'], function ($router) {
-    Route::get('', [PosyanduController::class, 'index']);
-    Route::get('{id}', [PosyanduController::class, 'show']);
-    Route::post('store', [PosyanduController::class, 'store']);
-    Route::delete('{id}', [PosyanduController::class, 'destroy']);
-    Route::post('reset_password', [PosyanduController::class, 'reset_password']);
+    Route::get('', [PosyanduController::class, 'index'])->middleware('role:kelurahan'); 
+    Route::get('{id}', [PosyanduController::class, 'show'])->middleware('role:kelurahan,posyandu');
+    Route::post('store', [PosyanduController::class, 'store'])->middleware('role:kelurahan');
+    Route::delete('{id}', [PosyanduController::class, 'destroy'])->middleware('role:kelurahan');
+    Route::post('reset_password', [PosyanduController::class, 'reset_password'])->middleware('role:kelurahan');
 });
 
 Route::group(['middleware' => 'api', 'prefix' => 'folder'], function ($router) {
-    Route::get('', [FolderController::class, 'index']);    
-    Route::get('{id}', [FolderController::class, 'show']);    
+    Route::get('', [FolderController::class, 'index'])->middleware('role:kelurahan,posyandu');
+    Route::get('{id}', [FolderController::class, 'show'])->middleware('role:kelurahan,posyandu');
+    Route::post('store', [FolderController::class, 'store'])->middleware('role:posyandu');
+    Route::get('based_posyandu/{posyandu_id}', [FolderController::class, 'based_posyandu'])->middleware('role:kelurahan,posyandu');
+    Route::get('{folder_id}/{is_terdata}', [FolderController::class, 'list_children'])->middleware('role:posyandu');
 });
 
 Route::group(['middleware' => 'api', 'prefix' => 'data'], function ($router) {
-    Route::get('based_folder/{folder_id}', [DataCollectionController::class, 'based_folder']);
-    Route::get('based_posyandu/{posyandu_id}', [DataCollectionController::class, 'based_posyandu']);
-    Route::get('based_children/{children_id}', [DataCollectionController::class, 'based_children']);
+    Route::get('based_folder/{folder_id}', [DataCollectionController::class, 'based_folder'])->middleware('role:kelurahan');
+    Route::get('based_posyandu/{posyandu_id}', [DataCollectionController::class, 'based_posyandu'])->middleware('role:kelurahan');
+    Route::get('based_children/{children_id}', [DataCollectionController::class, 'based_children'])->middleware('role:kelurahan');
+    Route::post('store', [DataCollectionController::class, 'store'])->middleware('role:posyandu');
+    Route::get('history/{children_id}', [DataCollectionController::class, 'history'])->middleware('role:posyandu');
+    Route::delete('{data_id}', [DataCollectionController::class, 'destroy'])->middleware('role:posyandu');
 });
 
 Route::group(['middleware' => 'api', 'prefix' => 'children'], function ($router) {
-    Route::get('{id}', [ChildrenController::class, 'show']);
+    Route::get('{id}', [ChildrenController::class, 'show'])->middleware('role:kelurahan,posyandu');
+    Route::post('store', [ChildrenController::class, 'store'])->middleware('role:posyandu');
+    Route::put('{id}', [ChildrenController::class, 'update'])->middleware('role:posyandu');
+    Route::delete('{id}', [ChildrenController::class, 'destroy'])->middleware('role:posyandu');
 });
